+++ CHANGES.rst
@@ -20,6 +20,7 @@ Development
 - Add ``SystemUptime`` sensor
 - Fix: Better explicitly initialize the ADC before reading it
 - Improve ``SystemBatteryLevel`` sensor. Thanks, `Dominik <https://github.com/ayoy>`_!
+- Make ``TelemetryTransportHTTP`` work again
 
 
 2019-06-07 0.3.0